+++ source/renderer/app/containers/wallet/AdaRedemptionPage.js
@@ -38,7 +38,7 @@ export default class AdaRedemptionPage extends Component<Props> {
     } = adaRedemption;
     const {
       chooseRedemptionType, setCertificate, setPassPhrase, setRedemptionCode, removeCertificate,
-      setEmail, setAdaPasscode, setAdaAmount, acceptRedemptionDisclaimer
+      setEmail, setAdaPasscode, setAdaAmount, setDecryptionKey, acceptRedemptionDisclaimer
     } = this.props.actions.ada.adaRedemption;
     const selectableWallets = wallets.all.map((w) => ({
       value: w.id, label: w.name
@@ -51,12 +51,20 @@ export default class AdaRedemptionPage extends Component<Props> {
     );
     const isCertificateSelected = adaRedemption.certificate !== null;
 
-    const showInputsForDecryptingForceVendedCertificate = isCertificateSelected &&
-      isCertificateEncrypted && redemptionType === ADA_REDEMPTION_TYPES.FORCE_VENDED;
-
-    const showPassPhraseWidget = isCertificateSelected && isCertificateEncrypted &&
-      redemptionType === ADA_REDEMPTION_TYPES.REGULAR ||
-      redemptionType === ADA_REDEMPTION_TYPES.PAPER_VENDED;
+    const showInputsForDecryptingForceVendedCertificate = (
+      isCertificateSelected && isCertificateEncrypted &&
+      redemptionType === ADA_REDEMPTION_TYPES.FORCE_VENDED
+    );
+    const showInputForDecryptionKey = (
+      isCertificateSelected && isCertificateEncrypted &&
+      redemptionType === ADA_REDEMPTION_TYPES.RECOVERY_FORCE_VENDED
+    );
+    const showPassPhraseWidget = redemptionType === ADA_REDEMPTION_TYPES.PAPER_VENDED || (
+      isCertificateSelected && isCertificateEncrypted && (
+        redemptionType === ADA_REDEMPTION_TYPES.REGULAR ||
+        redemptionType === ADA_REDEMPTION_TYPES.RECOVERY_REGULAR
+      )
+    );
 
     return (
       <Layout>
@@ -69,6 +77,7 @@ export default class AdaRedemptionPage extends Component<Props> {
           onEmailChanged={(email) => setEmail.trigger({ email })}
           onAdaAmountChanged={(adaAmount) => setAdaAmount.trigger({ adaAmount })}
           onAdaPasscodeChanged={(adaPasscode) => setAdaPasscode.trigger({ adaPasscode })}
+          onDecryptionKeyChanged={(decryptionKey) => setDecryptionKey.trigger({ decryptionKey })}
           onChooseRedemptionType={(choice) => {
             chooseRedemptionType.trigger({ redemptionType: choice });
           }}
@@ -91,6 +100,7 @@ export default class AdaRedemptionPage extends Component<Props> {
           showInputsForDecryptingForceVendedCertificate={
             showInputsForDecryptingForceVendedCertificate
           }
+          showInputForDecryptionKey={showInputForDecryptionKey}
           showPassPhraseWidget={showPassPhraseWidget}
           isRedemptionDisclaimerAccepted={environment.isMainnet() || isRedemptionDisclaimerAccepted}
           onAcceptRedemptionDisclaimer={() => acceptRedemptionDisclaimer.trigger()}